Handlers errors add

+++ backend/controllers/cards.js
@@ -1,25 +1,31 @@
 const Card = require('../models/card');
+const { BadRequestError, UnauthorizedError, ForbiddenError, NotFoundError, InternalServerError } = require('../middlewares/errors');
 
-module.exports.getCards = (req, res) => {
+module.exports.getCards = (req, res, next) => {
   Card.find({})
     .then((card) => res.send({ data: card }))
-    .catch(() => res.status(500).send({ message: 'На сервере произошла ошибка' }));
+    .catch(next);
 };
 
-module.exports.createCard = (req, res) => {
-  // console.log(req.user._id); // _id станет доступен
+module.exports.createCard = (req, res, next) => {
 
   const { name, link } = req.body;
   const owner = req.user._id;
   Card.create({ name, link, owner })
     .then((card) => res.send({ data: card }))
-    .catch(() => res.status(500).send({ message: 'На сервере произошла ошибка' }));
+    .catch(next)
 };
 
-module.exports.deleteCard = (req, res) => {
-  // console.log(req.params);
-  Card.findByIdAndRemove(req.params.cardId)
-    .then((card) => { if (card == null) res.status(404).send({ message: 'Карточка не существует' }); })
-    .then((card) => res.send({ data: card }))
-    .catch(() => res.status(500).send({ message: 'На сервере произошла ошибка' }));
+module.exports.deleteCard = (req, res, next) => {
+  Card.findById(req.params.cardId.toString())
+    .then((card) => {
+      if (card == null)
+      throw new NotFoundError( 'Карточка не существует' )
+        if (card.owner.toString() === req.user._id) {
+          Card.findByIdAndRemove(req.params.cardId)
+          .then((card) => {  return res.send({ data: card }) })
+          throw new InternalServerError( 'На сервере произошла ошибка' )
+      }
+  })
+  .catch(next)
 };
